fix(cds): [CDS 系统设置] env-loader 提前到 config 模块求值之前 + 空值覆盖语义
修复"GitHub App 配置 .cds.env 写好了但 /api/github/app 一直返回
configured:false"的幽灵 503 webhook bug。两条独立的失效路径同时显形：

Bug 1（ES module 导入顺序）：原版 loadCdsEnvFile 写在 index.ts 函数体里，
但 index.ts 顶部 `import { loadConfig } from './config.js'` 会先触发
config.ts 的 module-level 求值——`DEFAULT_CONFIG.githubApp =
resolveGitHubApp()` 在那一刻读 process.env，但此时 self-loader 还没跑！
GitHub App config 永远 undefined。

修法：把 loadCdsEnvFile 抽到 cds/src/load-env.ts 独立 module，文件顶
层调用一次。然后让 config.ts 顶部 side-effect import './load-env.js'。
ES module 规范保证 load-env.ts 在 DEFAULT_CONFIG 求值前就已执行完。
index.ts 的 import 也保留为冗余防御（即便有人调整 import 顺序也不破）。

Bug 2（self-update spawn 透传 stale 空值）：branches.ts self-update /
self-force-sync 端点 spawn 子进程时用 `env: { ...process.env }`。如果
父进程 env 里某 key 是空字符串（被 init 写过空值，或被某工具回填空），
子进程 process.env[key] === '' 就跳过了原 self-loader 的「undefined 才
覆盖」分支。

修法：self-loader 语义改为 `existing === undefined || existing === ''`
都用磁盘最新值覆盖。systemd / shell 显式注入的非空值仍然有最高优先级。

副作用：消除一直困扰运维的"明明 .cds.env 配好了但 CDS UI 还显示
未配置"问题，让 push 即部署的承诺真正落地。

测试：1141/1141 vitest 通过 + tsc 零错误。
